+++ qnx/src/ipc3x_dev/ti/syslink/family/common/vayu/vayuipu/vayucore0/VAYUIpuCore0Proc.c
@@ -1502,9 +1502,9 @@ VAYUIPUCORE0PROC_read (Processor_Handle   handle,
 #endif /* #if !defined(IPC_BUILD_OPTIMIZE) */
         procPtr8 = (UInt8 *) procAddr ;
         buffer = memcpy (buffer, procPtr8, *numBytes);
-        GT_assert (curTrace, (buffer != (UInt32) NULL));
+        GT_assert (curTrace, (buffer != NULL));
 #if !defined(IPC_BUILD_OPTIMIZE)
-        if (buffer == (UInt32) NULL) {
+        if (buffer == NULL) {
             /*! @retval PROCESSOR_E_FAIL Failed in memcpy */
             status = PROCESSOR_E_FAIL;
             GT_setFailureReason (curTrace,
